requirements.txt
requests
beautifulsoup4
pandas
gspread
oauth2client
google-api-python-client
psycopg2-binary